+++ modules/output/materials/materials_default.py
@@ -1,0 +1,3 @@
+
+def materials_standart():
+    return 0
